utils/melodyExtraction/requirements.txt
Keras==2.3.0
Librosa==0.7.0 
Cython==0.29.30
madmom==0.16.1
numpy==1.18.5
SciPy==1.1.0
pysndfx==0.3.6
tensorflow-gpu==1.15.5